% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_pack.R
\name{fun_pack}
\alias{fun_pack}
\title{Funkcja liczaca emisje. Na podstawie danych z EEA.}
\usage{
fun_pack(
  dane = input,
  kategoria = "Passenger Cars",
  paliwo = "Petrol",
  segment = "Mini",
  euro = "Euro 5",
  technologia = "GDI",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{dane}{\itemize{
\item data frame, sklada sie z 50 wierszy o kolumnach: (Nat, segment, paliwo, technologia)
\item nat, to liczba losowo generowana
\item segment, character w tej kolumnie losowane sa wartosci sposrod 4 dostepnych
parametrów: "Mini", "Small", "Medium", "Large-SUV-Executive"
\item paliwo, okresla sie rodzaj paliwa
\item technologia, okresla rodzaj technologii oczyszczania spalin
}}

\item{kategoria}{\itemize{
\item character okresla kategorie pojazdu
}}

\item{euro}{\itemize{
\item Europejski standard emisji spalin
}}

\item{mode}{\itemize{
\item zawiera rodzaj drogi
}}

\item{substancja}{\itemize{
\item wektor character, przyjmuje substancje dla ktorych chcemy szacowac emisje
aby zobaczyc liste dostepnych wartosci dla poszczegolnych parametrow nalezy uzyc
funkcji unique dla danych "wskazniki"
np. unique(wskazniki$Category)
}}
}
\value{
data frame z obliczona emisja
}
\description{
Przyjmuje nastpujepujace argumenty:
}
